Move types to separate files, change style of high-level interface
