+++ test/high.jl
@@ -67,5 +67,3 @@ myvar2[1,:,:] .= 0.0
 @test ncread(fn4, "myvar2") == myvar2
 @test nc_char2string(ncread(fn1, "vchar")) == xs
 @test_throws NetCDF.NetCDFException ncread("nonexistant file", "a var")
-
-ncclose()
